fix first working version of parser
