+++ include/shell.h
@@ -12,6 +12,21 @@
 # include <termios.h>
 # include "../lib/libft/libft.h"
 
+typedef enum s_bool
+{
+	false,
+	true,
+}	t_bool;
+
+struct s_state
+{
+	t_bool	pipe;
+	t_bool	option;
+	t_bool	command;
+	t_bool	redirect;
+	t_bool	last;
+};
+
 typedef struct s_command
 {
 	char	*path;
@@ -21,17 +36,14 @@ typedef struct s_command
 	int		out;
 }	t_command;
 
-typedef enum s_bool
-{
-	false,
-	true,
-}	t_bool;
-
-t_list	*lexing(char *str);
-void	tokenize_lst(t_list **lst);
-void	print_lst(t_list *lst);
-void	init_signalhandlers(void);
-t_list	*parse(t_list *lexed_args);
-int		quote_count(char *str);
+t_list			*lexing(char *str);
+void			tokenize_lst(t_list **lst);
+void			print_lst(t_list *lst);
+void			init_signalhandlers(void);
+t_list			*parse(t_list *lexed_args);
+struct s_state	init_state(void);
+int				quote_count(char *str);
+t_command		*create_cmd(char *command, char *options, int in, int out);
+void			reset_cmd(char **command, char **options, int *in, int *out);
 
 #endif
